+++ src/main/java/com/intranet/service/NotificationService.java
@@ -12,7 +12,7 @@ public interface NotificationService {
 	public Notification findBySender(User sender);
 	public void remove(Notification notify);
 	public void save(Notification notify);
-	public Notification findOne(Integer id);
+	public Notification findOne(Long id);
 	public Set<Notification> findByCreationAndType(Date date, String type);
 	public List<Date> findByTypeLastWeek(String type);
 }
